Updated pass-finalisation for common reducer
