+++ src/modules/tests/journal-data/common/index.ts
@@ -3,5 +3,3 @@ export * from './candidate';
 export * from './examiner';
 export * from './test-centre';
 export * from './test-slot-attributes';
-
-export { journalDataReducer } from './journal-data.reducar';
