+++ src/modules/tests/journal-data/index.ts
@@ -1,1 +1,2 @@
 export * from './common';
+export { journalDataReducer, getJournalData } from './common/journal-data.reducer';
