+++ src/modules/tests/pass-completion/cat-c/__tests__/pass-completion.reducer.cat-c.spec.ts
@@ -3,9 +3,11 @@ import {
   PassCertificateNumberChanged,
   ProvisionalLicenseReceived,
   ProvisionalLicenseNotReceived,
+} from '../../common/pass-completion.actions';
+import {
   Code78Present,
   Code78NotPresent,
-} from '../../pass-completion.actions';
+} from '../pass-completion.cat-c.actions';
 
 describe('pass completion reducer', () => {
   it('should put the pass certificate number into the state on pass certificate number changed action', () => {
