+++ src/modules/tests/pass-completion/cat-c/index.ts
@@ -1,0 +1,3 @@
+export * from './pass-completion.cat-c.actions';
+export * from './pass-completion.cat-c.selector';
+export { passCompletionCatCReducer } from './pass-completion.cat-c.reducer';
