+++ src/modules/tests/pass-completion/index.ts
@@ -1,0 +1,3 @@
+export * from './common/pass-completion.actions';
+export * from './common/pass-completion.selector';
+export { passCompletionReducer } from './common/pass-completion.reducer';
